MR-205: new admin profile, and profile edit and password routes changed
