+++ src/Components/Shared/Profile/index.jsx
@@ -67,16 +67,20 @@ const Profile = ({ user, activities }) => {
           )}
         </div>
         <div className={styles.profileActivitiesContainer}>
-          <div className={styles.profileActivitiesTitle}>
-            <p>{'Related activities:'}</p>
-          </div>
-          <div className={styles.profileActivitiesList}>
-            {activities.map((actv) => (
-              <div key={actv._id}>
-                <p>{actv.name}</p>
-              </div>
-            ))}
-          </div>
+          {sessionStorage.role === 'MEMBER' && (
+            <div className={styles.profileActivitiesTitle}>
+              <p>{'Related activities:'}</p>
+            </div>
+          )}
+          {sessionStorage.role === 'MEMBER' && (
+            <div className={styles.profileActivitiesList}>
+              {activities.map((actv) => (
+                <div key={actv._id}>
+                  <p>{actv.name}</p>
+                </div>
+              ))}
+            </div>
+          )}
         </div>
       </div>
       <div className={styles.profileButtonsContainer}>
